refactoring and action func validation for all kinds
