+++ bin/si-mcp-server/src/server.ts
@@ -29,8 +29,8 @@ import { templateListTool } from "./tools/templateList.ts";
 import { changeSetAbandonTool } from "./tools/changeSetAbandon.ts";
 import { changeSetForceApplyTool } from "./tools/changeSetForceApply.ts";
 
-// import { schemaUpsertTool } from "./tools/schemaUpsert.ts"; // uncomment for local development
-// import { funcUpsertTool } from "./tools/funcUpsert.ts"; // uncomment for local development
+// import { schemaCreateOrEditTool } from "./tools/schemaCreateOrEdit.ts"; // uncomment for local development
+// import { funcCreateOrEditTool } from "./tools/funcCreateOrEdit.ts"; // uncomment for local development
 
 export function createServer(): McpServer {
   const server = new McpServer({
@@ -67,8 +67,8 @@ export function createServer(): McpServer {
   importPrompt(server);
   discoverPrompt(server);
 
-  // schemaUpsertTool(server); // uncomment for local development
-  // funcUpsertTool(server); // uncomment for local development
+  // schemaCreateOrEditTool(server); // uncomment for local development
+  // funcCreateOrEditTool(server); // uncomment for local development
 
   return server;
 }
